+++ main.go
@@ -17,7 +17,9 @@ import (
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/awscluster"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/awscontrolplane"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/awsmachinedeployment"
+	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/cluster"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/g8scontrolplane"
+	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/machinedeployment"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/aws/networkpool"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/validator"
@@ -45,7 +47,17 @@ func main() {
 		log.Fatalf("Unable to create G8s Control Plane admitter: %v", err)
 	}
 
+	clusterMutator, err := cluster.NewMutator(config)
+	if err != nil {
+		panic(microerror.JSON(err))
+	}
+
 	g8scontrolplaneMutator, err := g8scontrolplane.NewMutator(config)
+	if err != nil {
+		panic(microerror.JSON(err))
+	}
+
+	machinedeploymentMutator, err := machinedeployment.NewMutator(config)
 	if err != nil {
 		panic(microerror.JSON(err))
 	}
@@ -81,7 +93,9 @@ func main() {
 	handler.Handle("/mutate/awscluster", mutator.Handler(awsclusterMutator))
 	handler.Handle("/mutate/awsmachinedeployment", mutator.Handler(awsmachinedeploymentMutator))
 	handler.Handle("/mutate/awscontrolplane", mutator.Handler(awscontrolplaneMutator))
+	handler.Handle("/mutate/cluster", mutator.Handler(clusterMutator))
 	handler.Handle("/mutate/g8scontrolplane", mutator.Handler(g8scontrolplaneMutator))
+	handler.Handle("/mutate/machinedeployment", mutator.Handler(machinedeploymentMutator))
 	handler.Handle("/validate/awscluster", validator.Handler(awsclusterValidator))
 	handler.Handle("/validate/awscontrolplane", validator.Handler(awscontrolplaneValidator))
 	handler.Handle("/validate/awsmachinedeployment", validator.Handler(awsmachinedeploymentValidator))