+++ pkg/aws/g8scontrolplane/mutate_g8scontrolplane_AZ_test.go
@@ -182,9 +182,10 @@ func getG8sControlPlaneRAWByte(replicaNum int, release string) ([]byte, error) {
 			Name:      controlPlaneName,
 			Namespace: controlPlaneNameSpace,
 			Labels: map[string]string{
-				"giantswarm.io/control-plane":   controlPlaneName,
-				"giantswarm.io/organization":    "giantswarm",
-				"release.giantswarm.io/version": release,
+				"giantswarm.io/control-plane":            controlPlaneName,
+				"cluster-operator.giantswarm.io/version": "1.2.3",
+				"giantswarm.io/organization":             "giantswarm",
+				"release.giantswarm.io/version":          release,
 			},
 		},
 		Spec: infrastructurev1alpha2.G8sControlPlaneSpec{