+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Default the Cluster Operator Version Label in `Cluster` from `Release` CR.
+- Default the AWS Operator Version Label in `AWSCluster` from `Release` CR.
+- Default the AWS Operator Version Label in `AWSControlplane`, `AWSMachinedeployment` Mutators and add generic label defaulting from AWSCluster CR.
+- Default the Cluster Operator Version Label in `G8sControlplane`, `Machinedeployment` Mutators and add generic label defaulting from cluster CR.
+- Default the Master attributes in the `AWSCluster` Mutator for pre-HA versions.
+- Default the Release Version Label and refactor the `G8sControlplane` and `AWSControlPlane` Mutators.
+- Default the Release Version Label in the `AWSCluster`, `MachineDeployment` and `AWSMachineDeployment` CRs based on the `Cluster`CR
+
 ## [2.2.3] - 2020-11-18
 
 - Set `400` status code in the validator response if a request is invalid.